# relu forward and backward with vectortype