@zhangting2020 2021-03-25
贴一下你测试机器上，bug前的log，以及现在的log。